Fixed a few warnings.
Rev: src/gc.c:1.126
Rev: src/interpret.h:1.65
Rev: src/modules/Image/blit.c:1.48
Rev: src/modules/Image/colortable_lookup.h:1.13
Rev: src/modules/Image/matrix.c:1.29
Rev: src/modules/Parser/html.c:1.115
Rev: src/multiset.c:1.27

+++ src/interpret.h
@@ -5,7 +5,7 @@
 \*/
 
 /*
- * $Id: interpret.h,v 1.64 2000/08/17 18:20:44 grubba Exp $
+ * $Id: interpret.h,v 1.65 2000/08/23 18:46:36 grubba Exp $
  */
 #ifndef INTERPRET_H
 #define INTERPRET_H
@@ -178,7 +178,7 @@ enum apply_type
 
 #define APPLY_MASTER(FUN,ARGS) \
 do{ \
-  static int fun_,master_cnt=0; \
+  static int fun_, master_cnt=0; \
   struct object *master_ob=master(); \
   if(master_cnt != master_ob->prog->id) \
   { \
@@ -190,7 +190,7 @@ do{ \
 
 #define SAFE_APPLY_MASTER(FUN,ARGS) \
 do{ \
-  static int fun_,master_cnt=0; \
+  static int fun_, master_cnt=0; \
   struct object *master_ob=master(); \
   if(master_cnt != master_ob->prog->id) \
   { \
@@ -257,7 +257,7 @@ static inline void strict_apply_svalue(struct svalue *sval, INT32 args)
 }
 #else /* !__ECL */
 #define apply_low(O,FUN,ARGS) \
-  mega_apply(APPLY_LOW, (ARGS), (void*)(O),(void*)(FUN))
+  mega_apply(APPLY_LOW, (ARGS), (void*)(O),(void*)(ptrdiff_t)(FUN))
 
 #define strict_apply_svalue(SVAL,ARGS) \
   mega_apply(APPLY_SVALUE, (ARGS), (void*)(SVAL),0)
